requirements.dev.txt
-r requirements.txt

django-stubs==4.2.7
mypy==1.9.0
pytest==8.1.1
pytest-django==4.11.1
ipython==8.22.2
types_Markdown==3.7.0.20241204
types-python-dateutil==2.9.0.20250809
